requirements.txt
Flask==3.1.1
Pillow==11.2.1
numpy==2.3.1